// admin/src/store/remote/roles/Roles.Types.ts
export type Module = {
	moduleId: number;
	module: string;
	details: string;
};

export type Role = {
	roleId: number;
	role: string;
	details: string;
	createdAt: Date;
	updatedAt: Date;
	modules?: Module[];
};

export type RoleDTO = {
	name: string;
  	details: string;
  	permissionIds: number[];
  }